\name{accl.coefs}
\alias{accl.coefs}
\title{accelerometer coefficients}
\usage{accl.coefs}
\description{\code{accl.coefs} contains accelerometer coefficients for an example created by Vladimir Maximov for the project GyroLib.
}
\format{
  This data frame contains 12 columns of data (accelerometer coefficients).
}
\source{
Vlad Maximov, 2012
Scalar Calibration of Vector accelerometers and magnetometers,
GyroLib documentation
}
\references{
Vlad Maximov, 2012
Scalar Calibration of Vector accelerometers and magnetometers,
GyroLib documentation
}
\author{Jose Gama}
\examples{
data(accl.coefs)
accl.coefs
}
\keyword{datasets}
