\name{magn.coefs}
\alias{magn.coefs}
\title{magnetometer coefficients}
\usage{magn.coefs}
\description{\code{magn.coefs} contains magnetometer coefficients for an example created by Vladimir Maximov for the project GyroLib.
}
\format{
  This data frame contains 12 columns of data (magnetometer coefficients).
}
\source{
Vlad Maximov, 2012
Scalar Calibration of Vector accelerometers and magnetometers,
GyroLib documentation
}
\references{
Vlad Maximov, 2012
Scalar Calibration of Vector accelerometers and magnetometers,
GyroLib documentation
}
\author{Jose Gama}
\examples{
data(magn.coefs)
magn.coefs
}
\keyword{datasets}
